% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoeffding.R
\name{hoeffR}
\alias{hoeffR}
\title{Efficiently Compute the Multivariate Hoeffding's R Statistic}
\usage{
hoeffR(X, Y, method = "auto")
}
\arguments{
\item{X}{an n-by-l matrx where n is the sample size and l is the number of
covariates in the first group.}

\item{Y}{an n-by-m matrx where n is the sample size and m is the number of
covariates in the second group.}

\item{method}{A string indicating which algorithm to use to compute the
       U-statistic. Currently there are three options
\itemize{
 \item{\code{"auto"} - Function attemps to automatically pick the fastest
 method for the given data.}
 \item{\code{"naive"} - A naive implementation which takes
 O(n^(4+l+m)) time.}
 \item{\code{"orth"} - A more complicated algorithm which takes
 O(n^(l+m)) time. Except in cases where l+m is large and n is small
 this algorithm tends to be significantly faster than the naive version.}
}}
}
\value{
the U-statistic for the data
}
\description{
Computes, from data, the U-statistic estimating Hoeffding's R, a measure
of dependence between random vectors X and Y.
}
