% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.R
\name{print.SymRCTest}
\alias{print.SymRCTest}
\title{Print Test Results}
\usage{
\method{print}{SymRCTest}(x, ...)
}
\arguments{
\item{x}{the SymRCTest object to be printed}

\item{...}{ignored.}
}
\description{
A simple print function for SymRCTest objects.
}
