% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateTauStar.R
\name{pTStar}
\alias{pTStar}
\title{Estimate the Multivariate Partial Tau Star Measure}
\usage{
pTStar(X, Y, method = "auto")
}
\arguments{
\item{X}{an n-by-l matrx where n is the sample size and l is the number of
covariates in the first group.}

\item{Y}{an n-by-m matrx where n is the sample size and m is the number of
covariates in the second group.}

\item{method}{A string indicating which algorithm to use to compute the
       U-statistic. Currently there are three options
\itemize{
 \item{\code{"auto"} - Function attemps to automatically pick the fastest
 method for the given data.}
 \item{\code{"def"} - Computes the U-statistic by definition, takes
  O(n^4) time.}
 \item{\code{"range-tree"} - A more complicated algorithm which takes
 O(n^2*log(n)^(2*(l+m)) time. While this is asymptotically faster than
 the naive algorithm it, in practice, can be slower due to
 large constant factors hidden by the asymptotic analysis.}
}}
}
\value{
the U-statistic for the data
}
\description{
Computes, from data, the U-statistic estimating the multivariate partial tau
star measure. This statistic estimates the dependence between two collections
of random variables X and Y.
}
