% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SymRC.R
\name{SymRC}
\alias{SymRC}
\title{SymRC package documentation.}
\description{
Estimates various Symmetric Rank Covariances (SRCs) using U-statistics.
Currently this packages provides implementations for quickly computing
Hoeffding's R and D statistics as well as two multivariate extensions of
the Bergsma/Dassios Sign Covariance tau*. See the following functions for
details:
\itemize{
 \item{\code{\link{pTStar}}}
 \item{\code{\link{jTStar}}}
 \item{\code{\link{hoeffD}}}
 \item{\code{\link{hoeffR}}}
}
}
\details{
To test whether or not two random vectors X and Y are dependent (using the
above U-statistics) see the function \code{\link{symRCTest}}.
}
