% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing.R
\name{symRCTest}
\alias{symRCTest}
\title{Independence Testing Using Symmetric Rank Covariances.}
\usage{
symRCTest(X, Y, measure = "hoeffD", resamples = 1000)
}
\arguments{
\item{X}{an n-by-l matrx where n is the sample size and l is the number of
covariates in the first group.}

\item{Y}{an n-by-m matrx where n is the sample size and m is the number of
covariates in the second group.}

\item{measure}{a string representing the SRC to use, can be one of
\itemize{
 \item{\code{\link{pTStar}}}
 \item{\code{\link{jTStar}}}
 \item{\code{\link{hoeffD}}}
 \item{\code{\link{hoeffR}}}
}}

\item{resamples}{the number of times permutation test iterations to use,
more iterations means a more stable p-value but long
computation time.}
}
\value{
an "SymRCTest" object (i.e. a list) recording the results of the
        independence test.
}
\description{
Tests the null hypothesis that two random vectors X and Y are independent
using symmetric rank covariances (SRCs). In particular, a permutation test is
performed using a U-statistic estimating a specified SRC.
}
\examples{
# Test whether X = (x1, x2) is independent from Y = (x1 * x2)^2 + (noise)
# Using the hoeffD measure.
set.seed(2)
n = 150
x1 = rnorm(n)
x2 = rnorm(n)
X = cbind(x1, x2)
Y = matrix((x1 * x2)^2 + rnorm(n))
symRCTest(X, Y, measure = "hoeffD")
}
