% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateTauStar.R
\name{pTStar}
\alias{pTStar}
\title{Estimate the Multivariate Partial Tau Star Measure}
\usage{
pTStar(X, Y, standardize = FALSE, method = "auto", approx = FALSE,
  approxControl = list(nreps = 10^4))
}
\arguments{
\item{X}{an n-by-l matrx where n is the sample size and l is the number of
covariates in the first group.}

\item{Y}{an n-by-m matrx where n is the sample size and m is the number of
covariates in the second group.}

\item{standardize}{if TRUE then this function returns a estimate of the
"standardized" version of the statistic which is bounded between -1
and 1. That is, letting the multivariate measure be denoted m, then
when \code{standardize = TRUE} this function returns an estimate of
m(X,Y) / sqrt(m(X,X) * m(Y,Y)). Estimates of m(X,X) and m(X,Y) are
done using settings defined below, if either m(X,X) or m(Y,Y) equal
0 then NA is returned and a warning is printed.}

\item{method}{A string indicating which algorithm to use to compute the
       U-statistic. Currently there are three options
\itemize{
 \item{\code{"auto"} - Function attemps to automatically pick the fastest
 method for the given data.}
 \item{\code{"def"} - Computes the U-statistic by definition, takes
  O(n^4) time.}
 \item{\code{"range-tree"} - A more complicated algorithm which takes
 O(n^2*log(n)^(2*(l+m)) time. While this is asymptotically faster than
 the naive algorithm it, in practice, can be slower due to
 large constant factors hidden by the asymptotic analysis.}
}
This argument is ignored if approx is TRUE.}

\item{approx}{if \code{approx} is \code{TRUE} then instead of computing the
u-statistic using all of the data we instead approximate it by
random sampling, see \code{approxControl} for how
to set the number of samples used.}

\item{approxControl}{a named list, should be of the form
\code{list(nsims = NSIMS)} if you would like to
approximate the u-statistic using \code{NSIMS} samples,
if instead you would like use as many samples as possible
within approximately some number of seconds, say
\code{NSECONDS}, then use a list of the form
\code{list(nseconds = NSECONDS)}.}
}
\value{
the (estimated if approx != 0) U-statistic for the data
}
\description{
Computes, from data, the U-statistic estimating the multivariate partial tau
star measure. This statistic estimates the dependence between two collections
of random variables X and Y. This U-statistic takes on
values between 0 and 2/3 when X and Y are 1-dimensional and between 0 and 2
when either X or Y are more than 1-dimensional.
}
\examples{
# Bivariate dependence
set.seed(1)
x = rnorm(30)
y = x^2 + rnorm(30)
pTStar(x, y)

# Trivariate dependence
x1 = rnorm(30)
x2 = rnorm(30)
y = (x1 * x2)^2 + rnorm(30)
pTStar(cbind(x1, x2), y)

# Approximate the above as well as we can in 5 seconds
pTStar(cbind(x1, x2), y, approx = T, approxControl = list(nseconds = 5))
}
